[fix] transaction 범위 변경 #25
