+++ Api/src/main/java/allchive/server/api/user/service/GetUserInfoUseCase.java
@@ -9,7 +9,9 @@ import allchive.server.domain.domains.category.adaptor.CategoryAdaptor;
 import allchive.server.domain.domains.category.domain.Category;
 import allchive.server.domain.domains.user.adaptor.UserAdaptor;
 import allchive.server.domain.domains.user.domain.User;
+import allchive.server.domain.domains.user.validator.UserValidator;
 import lombok.RequiredArgsConstructor;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -17,9 +19,12 @@ import java.util.List;
 @RequiredArgsConstructor
 public class GetUserInfoUseCase {
     private final UserAdaptor userAdaptor;
+    private final UserValidator userValidator;
 
+    @Transactional(readOnly = true)
     public GetUserInfoResponse execute() {
         Long userId = SecurityUtil.getCurrentUserId();
+        userValidator.validateUserStatusNormal(userId);
         User user = userAdaptor.queryUserById(userId);
         return GetUserInfoResponse.from(user);
     }
