+++ Api/src/main/java/allchive/server/api/user/service/GetUserProfileUseCase.java
@@ -8,7 +8,9 @@ import allchive.server.domain.domains.category.adaptor.CategoryAdaptor;
 import allchive.server.domain.domains.category.domain.Category;
 import allchive.server.domain.domains.user.adaptor.UserAdaptor;
 import allchive.server.domain.domains.user.domain.User;
+import allchive.server.domain.domains.user.validator.UserValidator;
 import lombok.RequiredArgsConstructor;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -18,9 +20,12 @@ public class GetUserProfileUseCase {
     private final CategoryAdaptor categoryAdaptor;
     private final UserAdaptor userAdaptor;
     private final UserMapper userMapper;
+    private final UserValidator userValidator;
 
+    @Transactional(readOnly = true)
     public GetUserProfileResponse execute() {
         Long userId = SecurityUtil.getCurrentUserId();
+        userValidator.validateUserStatusNormal(userId);
         User user = userAdaptor.queryUserById(userId);
         List<Category> categoryList = categoryAdaptor.findAllByUserId(userId);
         return userMapper.toGetUserProfileResponse(categoryList, user);
