+++ Api/src/main/java/allchive/server/api/user/service/UpdateUserInfoUseCase.java
@@ -8,17 +8,20 @@ import allchive.server.core.annotation.UseCase;
 import allchive.server.domain.domains.user.adaptor.UserAdaptor;
 import allchive.server.domain.domains.user.domain.User;
 import allchive.server.domain.domains.user.service.UserDomainService;
+import allchive.server.domain.domains.user.validator.UserValidator;
 import lombok.RequiredArgsConstructor;
+import org.springframework.transaction.annotation.Transactional;
 
 @UseCase
 @RequiredArgsConstructor
 public class UpdateUserInfoUseCase {
-    private final UserAdaptor userAdaptor;
     private final UserDomainService userDomainService;
+    private final UserValidator userValidator;
 
-
+    @Transactional
     public void execute(UpdateUserInfoRequest request) {
         Long userId = SecurityUtil.getCurrentUserId();
+        userValidator.validateUserStatusNormal(userId);
         String imgKey = UrlUtil.convertUrlToKey(request.getImgUrl());
         userDomainService.updateUserInfo(userId, request.getName(), request.getEmail(), request.getNickname(), imgKey);
     }
