+++ Domain/src/main/java/allchive/server/domain/domains/category/adaptor/CategoryAdaptor.java
@@ -59,7 +59,6 @@ public class CategoryAdaptor {
         return categoryRepository.queryCategoryExist(categoryId);
     }
 
-    @Transactional
     public List<Category> findAllByUserId(Long userId) {
         return categoryRepository.findAllByUserId(userId);
     }
